requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
pydantic>=2.7
python-dotenv>=1.0
requests>=2.32
tavily-python>=0.3.5
beautifulsoup4>=4.12
pypdf>=5.1

